add alert to index ts and refactor. Everything except AlertActions cause of cycling dependencies
