+++ src/shared/hooks/useAppDispatch.ts
@@ -1,3 +1,4 @@
+'use client'
 import type { AppDispatch } from '@/app/store';
 
 import { useDispatch } from 'react-redux';
