+++ src/shared/hooks/useAppSelector.ts
@@ -1,3 +1,4 @@
+'use client'
 import type { RootState } from '@/app/store';
 
 import { useSelector } from 'react-redux';
