+++ src/shared/hooks/useAppStore.ts
@@ -1,3 +1,4 @@
+'use client'
 import type { AppStore } from '@/app/store';
 
 import { useStore } from 'react-redux';
