+++ src/shared/ui/alert/index.ts
@@ -1,0 +1,3 @@
+export * from './lib/hooks/useAlert'
+export * from './types/types'
+export * from './ui/Alert'
