+++ src/shared/ui/alert/ui/Alert.module.scss
@@ -1,4 +1,4 @@
-@import 'src/app/styles/mixins';
+@import '../../../../app/styles/mixins';
 
 @keyframes visible {
   0% {
