+++ src/shared/ui/alert/ui/Alert.stories.tsx
@@ -1,9 +1,7 @@
 import type { Meta } from '@storybook/react';
 
 import StoreProvider from '@/app/StoreProvider';
-import useAlert from '@/shared/alert/lib/hooks/useAlert';
-import { Alert } from '@/shared/alert/ui/Alert';
-import { Button } from '@/shared/ui';
+import { Alert, Button, useAlert } from '@/shared/ui';
 
 const meta = {
   component: Alert,
