+++ src/shared/ui/alert/ui/Alert.tsx
@@ -1,11 +1,11 @@
 'use client';
 
 import CloseOutline from '@/../public/assets/components/CloseOutline';
-import { alertSelectors } from '@/shared/alert/model/alertSlice';
 import { useActions } from '@/shared/hooks/useActions';
 import { useAppSelector } from '@/shared/hooks/useAppSelector';
 import { Typography } from '@/shared/ui';
 import { Button } from '@/shared/ui';
+import { alertSelectors } from '@/shared/ui/alert/model/alertSlice';
 import { clsx } from 'clsx';
 
 import s from './Alert.module.scss';
