+++ src/shared/ui/index.ts
@@ -1,3 +1,4 @@
+export * from './alert';
 export * from './backBtn/BackBtn';
 export * from './button/Button';
 export * from './card/Card';
